Replace deprecated tpl_action by \dokuwiki\Menu
This commit introduces two changes which either resulted error or
deprecation log events on a current "2023-04-04a "Jack Jackrum""
DokuWiki installation running on PHP 8.2.

First, the use of tpl_action became deprecated and should be replaced by
\dokuwiki\Menu[0]. This change went along the documentation.

Second, the $_SERVER['REMOTE_USER'] is missing on my installation and
does not seem to be necessary. For example, the tpl_userinfo() function
has no effect if no user is logged in[1].

[0] https://www.dokuwiki.org/devel:menus
[1] https://codesearch.dokuwiki.org/xref/dokuwiki/inc/template.php?r=&mo=19006&fi=674#888